refactoring location stuff
