+++ tests/base/test_goto.lua
@@ -35,7 +35,7 @@ T['can goto given location'] = function()
   })
   helpers.childWaitForFinishedStatus(child)
 
-  child.type_keys('<esc>10G')
+  child.type_keys('<esc>10G0')
   child.type_keys('<esc>' .. keymaps.gotoLocation.n)
   helpers.childWaitForScreenshotText(child, '3,13')
   helpers.childExpectScreenshot(child)
@@ -86,7 +86,7 @@ T['can goto given location when only window in tabpage'] = function()
   })
   helpers.childWaitForFinishedStatus(child)
 
-  child.type_keys('<esc>10G')
+  child.type_keys('<esc>10G0')
   child.type_keys('<esc>' .. keymaps.gotoLocation.n)
   helpers.childWaitForScreenshotText(child, '3,13')
   helpers.childExpectScreenshot(child)
