+++ tests/screenshots/tests-astgrep-test_goto.lua---can-open-a-given-location
@@ -21,5 +21,5 @@
 20|                                        │                                       
 21|                                        │                                       
 22|                                        │                                       
-23|file2.ts                             3,7 Grug FAR - 1: grug.$A               9,2
+23|file2.ts                             3,7 Grug FAR - 1: grug.$A               9,3
 24|
