+++ tests/screenshots/tests-base-test_goto.lua---can-goto-given-location
@@ -21,5 +21,5 @@
 20|                                        │                                       
 21|                                        │file2.doc                              
 22|                                        │  2       grug talks and grug drink [3]
-23|file1.txt                           3,13 Grug FAR - 1: grug                 10,2
+23|file1.txt                           3,13 Grug FAR - 1: grug                 10,3
 24|
